Resolve type parameters consecutively
Previously, we were resolving all type parameter constraints before
resolving the type parameters themselves. This resulted in unresolved
types in cases where a later type parameter relied on a previous type
parameter.
